requirements.txt
python-dotenv==1.0.0
openai==1.7.0
Flask==3.0.0